Tally isPcr output by lineage and build reports
This is an initial attempt to add a command line API
to two scripts producing reports from the output of
Jim Kent's isPCR tool vs taxonomic lineage.

Calling these v0.9.0 as expect to tweak the command
line API still, especially for the reporting script.

This was written as part of an in-silico PCR primer
evaluation project.