Apply root filter and cut on loading
This makes the multiple products warning more
useful, and is much faster.

The original script loaded everything once, and
then looped over several roots making multiple
reports. This version with a command line API
is limited to reporting on a single root.
Could refactor more to apply the level cut
on loading etc as well.