Demo of calling snakemake via API with a progress bar
Currently the snakemake API doesn't have any obvious way
to get callbacks or an iterator approach to running a
workflow which would allow direct updates to a progress
bar. Improvements to their logging system may allow this?

Instead, this demonstrate running snakemake on a subprocess,
and monitoring the creation of the expected output files
as a proxy to update a progress bar. This works, but would
put some additional load on the file system.